api/requirements.txt
fastapi
uvicorn[standard]
openai
sqlalchemy
pydantic
python-dotenv
